+++ models/KemasanModel.js
@@ -13,7 +13,7 @@ const KemasanModel = db.define(
       },
     },
     harga: {
-      type: Sequelize.DataTypes.INTEGER,
+      type: Sequelize.DataTypes.DECIMAL(10, 3),
       allowNull: false,
       validate: {
         notEmpty: true,
